Move all the .h headers to .hpp headers
